refactor: yaml 파일 Long 타입 수정 및 코드 수정

+++ backend/src/main/kotlin/com/woowahan/campus/announcement/domain/Announcement.kt
@@ -13,7 +13,7 @@ class Announcement(
     var content: Content,
     @Embedded
     var author: Author,
-    val slackChannelId: Int,
+    val slackChannelId: Long,
     id: Long = 0L,
 ) : BaseRootEntity<Announcement>(id) {
